requirements.txt
python-dotenv==1.0.1
typing_extensions==4.12.2
